Add Test Util For Common Testing Patterns
Add the test util for common testing patterns that happen throughout the
application, such as using a custom render so that each time has to use
the same wrapper, thus reducing the test's complexity. Furthermore, it
makes the test more manageable and allows the developer to create more
of them.