Add Convert API to App Cars Helper Function
Add Convert API to App Cars Helper Function as other non-state
management will be using the App Cars object as well, so both of them
need to convert the API Cars object o the App's version of Cars objects.
To allow for better developer experience when working on the
application.