Update Search Value in Search Value Type
Update the Search Value to use an object as its value so that the matter
can continuously be updated without having to search for a string value
in the search value and parse it to update the value with the new value.
Later, when we want to search with the Search Value, we will convert it
to a string that the Realm filter can use to search the database.