Add Dialog & Color Picker Slider to Search Comp
Add dialogue and Color Picker Slider to the Search component since the
dialogue will have all the filter options that the user can select,
including clearing the search. The colour picker is a slider as there
are many different colours a car can have, and we should give as many
colours as possible since the user may select a rare car colour that we
may realize in a dropdown.