Move Database Logic to `useDatabase` Hook
Move the database logic to the `useDatabase` hook so that other parts of
the application use the database. Furthermore, this can help with
testing as the developer can isolate the hook from the AppBody, thus
seeing what the hook will return with little effect. The AppBody is also
a lot clearer as the database's logic is in a different area.