Add Fast Image Package
Add the Fast Image package to render and cache images since the inherent
issue with having a list of images is also called the server for each
image. Fast Image is supposed to cache those images where it can load
those if the user renders the list again, saving a trip to the server to
get the newest images.

Ref:

https://www.npmjs.com/package/expo-react-native-fast-image