Remove `.env.example` to Hardcode URLs
Remove `.env.example` to hardcode URLs as Snack Expo has issues with
using environment variables, causing the App not to work in Snack Expo.
So typical, this would be in `.env.example` and be copied to any
environment so that the environment variables can be picked up by the
deployment process and filled in with which environment the deployment
is happening. It looks like Expo does not have it, or the documentation
is wrong.